+++ src/main/kotlin/me/modmuss50/mpp/networking/MultipartBodyBuilder.kt
@@ -1,4 +1,4 @@
-package me.modmuss50.mpp
+package me.modmuss50.mpp.networking
 
 import java.io.File
 import java.net.http.HttpRequest
@@ -82,7 +82,8 @@ class MultipartBodyBuilder {
             val result = mutableListOf<ByteArray>()
             val contentType = mediaType ?: Files.probeContentType(path) ?: "application/octet-stream"
 
-            result.add("Content-Disposition: form-data; name=\"$name\"; filename=\"$filename\"\r\n".toByteArray(StandardCharsets.UTF_8))
+            result.add("Content-Disposition: form-data; name=\"$name\"; filename=\"$filename\"\r\n".toByteArray(
+                StandardCharsets.UTF_8))
             result.add("Content-Type: $contentType\r\n\r\n".toByteArray(StandardCharsets.UTF_8))
             result.add(Files.readAllBytes(path))
             result.add("\r\n".toByteArray(StandardCharsets.UTF_8))
