+++ src/main/kotlin/me/modmuss50/mpp/platforms/curseforge/CurseforgeApi.kt
@@ -3,7 +3,7 @@ package me.modmuss50.mpp.platforms.curseforge
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.encodeToString
-import me.modmuss50.mpp.MultipartBodyBuilder
+import me.modmuss50.mpp.networking.MultipartBodyBuilder
 import me.modmuss50.mpp.PlatformDependency
 import me.modmuss50.mpp.platforms.HttpClients
 import java.nio.file.Path
